refactor: rename for plural
