+++ src/observatory/src/api/http.rs
@@ -1,4 +1,4 @@
-use crate::certificate::http::response::transform_certificate_response as transform_certificate;
+use crate::certificates::http::response::transform_certificate_response as transform_certificate;
 use crate::notification::http::response::transform_response;
 use ic_cdk::management_canister::{HttpRequestResult, TransformArgs};
 use ic_cdk_macros::query;
